Crate problem030

Problem 30 (Digit fifth powers)

Solution detail

We can find the solutions to this problem using a backtracking-style algorithm, building up the solutions from left-to-right one digit at a time. Note that since 7 × 95 has only 6 digits, there cannot be any solutions with more than 6 digits. This means that our search tree only has depth 6.

We can also prune the search tree before reaching the bottom by calculating the maximum and minimum values that can be obtained by by the sum of fifth powers after extending the number to length 6, and seeing if the number itself can ever fall within that range.

For example, no matter what three digits we add to the end of 211, the sum of fifth powers can never be larger than 25 + 1 + 1 + 95 + 95 + 95, which is equal to 177181. On the other hand, the number itself after adding three digits will always be at least 211000, so there cannot possibly be a solution.

Finally, remember to subtract 1 at the end, since it is not counted as a solution.

Structs

PowerSumTree
The information that is held about the current state during the tree search.

PowerSumTreeStep
A description of a step that can be taken in the search tree.

Constants

FIFTH_POWER
Fifth powers of single digits.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the sum of the numbers which are equal to the sum of the fifth powers of their digits.